pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "changetrail"
version = "0.1.0"
description = "Unified change timeline for incident context"
readme = "README.md"
license = {text = "MIT"}
authors = [{name = "cvemula1"}]
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy>=2.0.35",
    "asyncpg>=0.30.0",
    "psycopg2-binary>=2.9.9",
    "alembic>=1.13.2",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "httpx>=0.27.0",
    "kubernetes>=31.0.0",
    "python-dateutil>=2.9.0",
    "apscheduler>=3.10.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.6.0",
]

[project.urls]
Homepage = "https://github.com/cvemula1/ChangeTrail"
Repository = "https://github.com/cvemula1/ChangeTrail"
Issues = "https://github.com/cvemula1/ChangeTrail/issues"

[project.scripts]
changetrail = "changetrail.cli:main"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
